New classes to encapsulate the scene and animation
The `Coordinate` class is the most granular of the classes added by these changes. A coordinate is simply a generic list of values; the concrete subclass `Joint` entails a named part of a character's rig. `Character`, which subclasses `Thing`, contains a list of joints. In similar way, `Frame` contains a list of things (characters) and `Animation` contains a list of frames, paired with a `Timeline`.

Note that, with this design, the animation indirectly contains a list of things and each of those thing contains a specific configuration of coordinates. From that perspective, this design represents the animation more as baked data than as an animated character.